Refactor query_handler.rs into modular directory structure
Split the 1,960-line query_handler.rs into a query_handler/ module directory
with focused submodules for better maintainability:

- types.rs: All snapshot types, enums, and the UtxoQueryProvider trait
- utxo.rs: UTxO query handlers (tags 6, 7, 15)
- stake.rs: Stake/delegation query handlers (tags 10, 16, 17, 19, 20, 21, 22)
- governance.rs: Governance query handlers (tags 23, 24, 25, 26, 27, 28)
- protocol.rs: Protocol param, genesis, and reward handlers (tags 2, 3, 4, 5, 11, 29)
- mod.rs: QueryHandler struct, CBOR dispatch logic, re-exports, and tests

Pure refactoring with no behavioral changes. All public items re-exported
for backwards compatibility.

Closes #30
